closes #1, closes #2,  closes #3; 1 Missing `ucid` column. Solved. The series ucid was as an index became a column.
2 Strange variable order
Modify the order of the series so that the table starts with [pos, t_frame, ucid, cellID].
I modify indexes to maintain compatibility with rCell. Change multi index ucid and t_frame for numeric indices.

3 Modify fluorescence variable names.
I modify the name of the XFP fluorescence variables by xfp to maintain consistency with the rest of the variables.